# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Metaframer headless API",
    "version": "1.1.0",
    "description": "ERPNext is the source of truth; no mock fallback. Static admin/storefront frontends run on GitHub Pages and call the API-only server. AUTH_MODE=bearer uses an encrypted opaque 8-hour admin session; cookie mode remains the local default. Anonymous public catalog routes expose only enabled sales Items in the configured leaf Item Group. Internal catalog routes still require a separate server secret. Non-allowlisted browser Origins are rejected."
  },
  "servers": [
    {
      "url": "https://headless-api.metaframer.net",
      "description": "API-only production server; frontends are on https://metafrappe.github.io"
    },
    {
      "url": "http://localhost:4300",
      "description": "Local development; cookie authentication by default"
    }
  ],
  "security": [
    {
      "AdminBearer": []
    },
    {
      "AdminSession": []
    }
  ],
  "paths": {
    "/api/v1/auth/login": {
      "post": {
        "summary": "Authenticate against Frappe; no local password database",
        "security": [],
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Must exactly match APP_ORIGIN or ALLOWED_ORIGINS; the browser sets this automatically. Pages Origin is https://metafrappe.github.io, without repository path."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "username",
                  "password"
                ],
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 320
                  },
                  "password": {
                    "type": "string",
                    "format": "password",
                    "writeOnly": true,
                    "minLength": 1,
                    "maxLength": 1024
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "502": {
            "$ref": "#/components/responses/Error502"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "504": {
            "$ref": "#/components/responses/Error504"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "description": "Requires an exact allowlisted browser Origin. No local CSRF is required at login. Bearer mode returns accessToken without Set-Cookie; cookie mode returns Session plus encrypted HttpOnly cookie."
      },
      "options": {
        "summary": "Exact-origin CORS preflight",
        "security": [],
        "description": "204 validates allowed Origin/method/header names only; it does not prove the requested operation exists or authorize it. Missing/unlisted Origin or unlisted method/header returns 403.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorsOrigin"
          },
          {
            "$ref": "#/components/parameters/CorsMethod"
          },
          {
            "$ref": "#/components/parameters/CorsHeaders"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/CorsPreflight"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          }
        }
      }
    },
    "/api/v1/auth/session": {
      "get": {
        "summary": "Validate upstream session",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            },
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "502": {
            "$ref": "#/components/responses/Error502"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "504": {
            "$ref": "#/components/responses/Error504"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "description": "Validates upstream identity on every request; returns user and local csrfToken, not a replacement accessToken."
      },
      "options": {
        "summary": "Exact-origin CORS preflight",
        "security": [],
        "description": "204 validates allowed Origin/method/header names only; it does not prove the requested operation exists or authorize it. Missing/unlisted Origin or unlisted method/header returns 403.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorsOrigin"
          },
          {
            "$ref": "#/components/parameters/CorsMethod"
          },
          {
            "$ref": "#/components/parameters/CorsHeaders"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/CorsPreflight"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          }
        }
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "summary": "Invalidate the upstream Frappe session",
        "parameters": [
          {
            "name": "Origin",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uri"
            },
            "description": "Must exactly match APP_ORIGIN or ALLOWED_ORIGINS; the browser sets this automatically. Pages Origin is https://metafrappe.github.io, without repository path."
          },
          {
            "name": "X-CSRF-Token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Local csrfToken returned by login/session; not the upstream Frappe CSRF token."
          }
        ],
        "responses": {
          "204": {
            "description": "Logged out",
            "headers": {
              "X-Request-ID": {
                "$ref": "#/components/headers/RequestId"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Error401"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          },
          "404": {
            "$ref": "#/components/responses/Error404"
          },
          "409": {
            "$ref": "#/components/responses/Error409"
          },
          "422": {
            "$ref": "#/components/responses/Error422"
          },
          "429": {
            "$ref": "#/components/responses/Error429"
          },
          "502": {
            "$ref": "#/components/responses/Error502"
          },
          "503": {
            "$ref": "#/components/responses/Error503"
          },
          "504": {
            "$ref": "#/components/responses/Error504"
          },
          "400": {
            "$ref": "#/components/responses/Error400"
          },
          "500": {
            "$ref": "#/components/responses/Error500"
          }
        },
        "description": "204 confirms upstream logout succeeded. Cookie mode also clears its cookie. A failed upstream logout returns an error: clearing a local bearer token alone does not revoke a copied stateless token."
      },
      "options": {
        "summary": "Exact-origin CORS preflight",
        "security": [],
        "description": "204 validates allowed Origin/method/header names only; it does not prove the requested operation exists or authorize it. Missing/unlisted Origin or unlisted method/header returns 403.",
        "parameters": [
          {
            "$ref": "#/components/parameters/CorsOrigin"
          },
          {
            "$ref": "#/components/parameters/CorsMethod"
          },
          {
            "$ref": "#/components/parameters/CorsHeaders"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/CorsPreflight"
          },
          "403": {
            "$ref": "#/components/responses/Error403"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mf_admin_session",
        "description": "AUTH_MODE=cookie only (default). Encrypted HttpOnly SameSite=Lax cookie, Secure in production. A deployment accepts its configured mode, not both modes simultaneously."
      },
      "CatalogKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "Server-to-server CATALOG_SHARED_SECRET; not the admin accessToken or Frappe API token. Never embed in Pages/browser code."
      },
      "AdminBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque iron-session seal (not JWT)",
        "description": "AUTH_MODE=bearer only. Encrypted 8-hour accessToken from login. Send in Authorization only; query tokens and cookie fallback are not accepted. Every protected request revalidates the upstream Frappe user. Writes additionally require allowed Origin and X-CSRF-Token."
      }
    },
    "schemas": {
...
      "LoginResponse": {
        "description": "Response depends on deployment AUTH_MODE; cookie mode also sets mf_admin_session.",
        "oneOf": [
          {
            "$ref": "#/components/schemas/BearerLogin"
          },
          {
            "$ref": "#/components/schemas/Session"
          }
        ]
      },
...
    },
    "headers": {
      "RequestId": {
        "description": "Request correlation identifier.",
        "schema": {
          "type": "string"
        }
      },
      "NoStore": {
        "schema": {
          "type": "string",
          "const": "no-store"
        }
      }
    },
    "responses": {
      "Error401": {
        "description": "Authentication required",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          },
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        }
      },
      "Error403": {
        "description": "Origin, requested CORS method/header, CSRF or Frappe permission denied",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          },
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        }
      },
      "Error404": {
        "description": "Not found",
        "content": {
          "application/problem+json": {
            "schema": {
              "$ref": "#/components/schemas/Problem"
            }
          }
        },
        "headers": {
          "X-Request-ID": {
            "$ref": "#/components/headers/RequestId"
          },
          "Cache-Control": {
            "$ref": "#/components/headers/NoStore"
          }
        }
      },
...
    },
...
  },
...
}